requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.29.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
aiogram==3.2.0
python-multipart==0.0.6
aiohttp==3.9.1
aiofiles==23.2.1
Pillow==10.1.0
Jinja2==3.1.4
